Refactor NettyTransport#getLocalAddress()
